Put user interaction from XmlFormIO/FormIO to the caller.
No *IO class should have direct access to the UI.

New issue
Summary: Form update infos should have a struct
When e.g. FormManager calls FormIO.checkFormUpdates() and later
FormIO.updateForms(), it asks the user in between that there are forms
to update. As GUI/BusinessLogic separation this was moved into the
FormManager. But he has to get some information about what forms are
going to be updated and which data has changed.

There would be need of a struct/list with infos of
GenericUpdateInformation and XmlFormName.

Eric, do you have an idea here?