fmf_admin:UPDATE privilege on mysql & GRANT OPTION
FMF default super admin needs to have UPDATE privilege on mysql database as a
whole and not only on mysql.user in order to change other MySQL users passwords
with SET PASSWORD FOR.
It also needs GRANT OPTION to pass on this privilege to other FMF admins.
This will help to find a fix for GitHub issue #47 and should be generalized to
code used to create new users.